Update steam-ws-download
Fixes a bug where the argument processor will try to continue after all arguments have already been processed, resulting in repeated processing of nothing, and errors related to trying to shift when there's nothing to shift. This is caused by the fact that it shifts twice when processing arguments with a parameter such as "--user <name>". The solution was to check before shifting, and then break the loop if there's nothing left.